TensorProductBsplines

Index

IgaBase.QuadratureRule

TensorProductBsplines.TPEvaluationCache

TensorProductBsplines.TensorProductBspline

IgaBase.QuadratureRule — Method

QuadratureRule(Q::TensorProduct{Dim,<:AbstractQuadrule{1}})

Generate a tensor product quadrature rule where the quadrature points are stored as a CartesianProduct and the weights as a KroneckerProduct vector allowing for easy evaluation and numerical integration based on Kronecker products.

TensorProductBsplines.TPEvaluationCache — Type

TPEvaluationCache{Dim}

Type that caches spline basis functions as a KroneckerProduct{Dim} and a grid of points as a CartesianProduct{Dim}. The type is mutable such that the cached variables can be updated at runtime.

TensorProductBsplines.TensorProductBspline — Type

TensorProductBspline{Dim,Ders,T}(space, coeffs)
TensorProductBspline{Dim,Ders,T}(space; codim=1)

Construct a Dim dimensional TensorProductBspline patch that is mapped to Codim dimensional Euclidean space.